GPII-1534: Extract NexusTestUtils.js
