+++ tests/ReadDefaultsTests.js
@@ -2,47 +2,18 @@
 
 var fluid = require("infusion"),
     gpii = fluid.registerNamespace("gpii"),
-    jqUnit = fluid.require("node-jqunit"),
     kettle = fluid.registerNamespace("kettle"),
     path = require("path"),
     configPath = path.resolve(__dirname, "../configs");
 
 require("kettle");
 require("../src/Nexus.js");
+require("../src/test/NexusTestUtils.js");
 
 kettle.loadTestingSupport();
 
 fluid.defaults("gpii.tests.nexus.testGrade", {
     gradeNames: ["fluid.component"]
-});
-
-gpii.tests.nexus.verifyReadDefaultsResponse = function (body, request, expectedGradeNames) {
-    // TODO: Switch over to kettle.test.assertJSONResponse
-    var responseGradeSpec = JSON.parse(body);
-    var response = request.nativeResponse;
-    jqUnit.assertEquals("Response has 200 status code", 200, response.statusCode);
-    jqUnit.assertTrue("Response has JSON content-type",
-                      response.headers["content-type"].indexOf("application/json") === 0);
-    fluid.each(expectedGradeNames, function (gradeName) {
-        jqUnit.assertTrue("Read Defaults response has grade " + gradeName,
-                          fluid.hasGrade(responseGradeSpec, gradeName));
-    });
-};
-
-fluid.defaults("gpii.test.nexus.testCaseHolder", {
-    gradeNames: "kettle.test.testCaseHolder",
-    components: {
-        readDefaultsRequest: {
-            type: "kettle.test.request.http",
-            options: {
-                path: "/defaults/%gradeName",
-                port: 8081,
-                termMap: {
-                    gradeName: "{tests}.options.testGradeName"
-                }
-            }
-        }
-    }
 });
 
 gpii.tests.nexus.testDefs = [
@@ -61,7 +32,7 @@ gpii.tests.nexus.testDefs = [
             },
             {
                 event: "{readDefaultsRequest}.events.onComplete",
-                listener: "gpii.tests.nexus.verifyReadDefaultsResponse",
+                listener: "gpii.test.nexus.verifyReadDefaultsResponse",
                 args: ["{arguments}.0", "{readDefaultsRequest}", ["fluid.component", "gpii.tests.nexus.testGrade"]]
             }
         ]
